# docs/openapi.yaml
openapi: 3.1.0
info:
  title: JavaCraft Platform API
  version: 1.0.0
  description: |
    Catalog, identity, and progress endpoints are implemented. Submission and execution
    operations describe a later security-gated milestone and are not enabled.
servers:
  - url: http://localhost:8080/api/v1
tags:
  - name: Catalog
  - name: Identity
  - name: Learning
  - name: Execution
paths:
  /catalog:
    get:
      tags: [Catalog]
      summary: Get the published learning catalog
      operationId: getCatalog
      responses:
        "200":
          description: Database-backed public catalog
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Catalog"
  /tutorials:
    get:
      tags: [Catalog]
      summary: List published tutorials
      operationId: listTutorials
      responses:
        "200":
          description: Published tutorial summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TutorialSummary"
  /tutorials/{slug}:
    get:
      tags: [Catalog]
      summary: Get a tutorial and its sections
      operationId: getTutorial
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Tutorial content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tutorial"
        "404":
          $ref: "#/components/responses/NotFound"
  /challenges/{slug}:
    get:
      tags: [Catalog]
      summary: Get public challenge instructions (never hidden tests)
      operationId: getChallenge
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Public challenge details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
        "404":
          $ref: "#/components/responses/NotFound"
  /auth/csrf:
    get:
      tags: [Identity]
      summary: Obtain the CSRF token for state-changing operations
      operationId: getCsrfToken
      responses:
        "200":
          description: Token also stored in the XSRF-TOKEN cookie
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token:
                    type: string
  /auth/register:
    post:
      tags: [Identity]
      summary: Register a learner and create a session
      operationId: register
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created; issues an HttpOnly session cookie
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Learner"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          description: An account with the email already exists
  /auth/login:
    post:
      tags: [Identity]
      summary: Create a learner session
      operationId: login
      security:
        - csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated learner; issues an HttpOnly session cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Learner"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      tags: [Identity]
      summary: Get the current learner
      operationId: getCurrentLearner
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current learner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Learner"
        "401":
          description: Authentication required
  /auth/logout:
    post:
      tags: [Identity]
      summary: Revoke the current session
      operationId: logout
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        "204":
          description: Session revoked and cookie cleared
  /me/progress:
    get:
      tags: [Learning]
      summary: Get the authenticated learner's durable progress
      operationId: getProgress
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Track, tutorial completion, and skill evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnerProgress"
        "401":
          description: Authentication required
  /me/tutorials/{slug}/progress:
    put:
      tags: [Learning]
      summary: Save tutorial progress
      operationId: updateTutorialProgress
      security:
        - cookieAuth: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/Slug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [IN_PROGRESS, COMPLETED]
      responses:
        "200":
          description: Updated learner progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearnerProgress"
        "401":
          description: Authentication required
        "404":
          $ref: "#/components/responses/NotFound"
  /submissions:
    post:
      tags: [Execution]
      summary: Queue a submission for isolated execution (planned; disabled)
      operationId: createSubmission
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        "202":
          description: Accepted only when the isolated worker is implemented
        "503":
          description: Execution service is unavailable
  /executions/{executionId}:
    get:
      tags: [Execution]
      summary: Get sanitized execution results (planned; disabled)
      operationId: getExecution
      security:
        - cookieAuth: []
      parameters:
        - name: executionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Execution status and sanitized aggregate test results
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JAVACRAFT_SESSION
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
  responses:
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Problem:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      required: [title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
    Catalog:
      type: object
      required: [tutorials, challenge]
      properties:
        tutorials:
          type: array
          items:
            $ref: "#/components/schemas/TutorialSummary"
        challenge:
          $ref: "#/components/schemas/Challenge"
    TutorialSummary:
      type: object
      required: [slug, title, description, durationMinutes]
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
    Tutorial:
      type: object
      required: [slug, title, description, durationMinutes, sections]
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        sections:
          type: array
          items:
            type: object
            required: [title, body, exampleCode]
            properties:
              title:
                type: string
              body:
                type: string
              exampleCode:
                type: string
    Challenge:
      type: object
      required: [slug, title, level, category, description, requirements, starterCode]
      properties:
        slug:
          type: string
        title:
          type: string
        level:
          type: string
        category:
          type: string
        description:
          type: string
        requirements:
          type: array
          items:
            type: string
        starterCode:
          type: string
        skills:
          type: array
          items:
            type: string
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 128
        displayName:
          type: string
          minLength: 1
          maxLength: 80
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          maxLength: 128
    Learner:
      type: object
      required: [id, email, displayName]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
    LearnerProgress:
      type: object
      required: [currentTrack, completedTutorials, totalTutorials, progressPercent, tutorials, skills]
      properties:
        currentTrack:
          type: string
        completedTutorials:
          type: integer
        totalTutorials:
          type: integer
        progressPercent:
          type: integer
          minimum: 0
          maximum: 100
        tutorials:
          type: array
          items:
            type: object
            required: [slug, status]
            properties:
              slug:
                type: string
              status:
                type: string
                enum: [NOT_STARTED, IN_PROGRESS, COMPLETED]
        skills:
          type: array
          items:
            type: object
            required: [slug, name, proficiency]
            properties:
              slug:
                type: string
              name:
                type: string
              proficiency:
                type: integer
                minimum: 0
                maximum: 100